add: add new conftest to fim
